fix correctly localizing cert-manager images
