+++ e2e-cli/suite_test.go
@@ -65,7 +65,7 @@ func TestMain(m *testing.M) {
 
 	testEnv.Setup(
 		envfuncs.CreateKindCluster(kindClusterName),
-		envfuncs.CreateNamespace("ocm-system"),
+		envfuncs.CreateNamespace(env.DefaultOCMNamespace),
 		envfuncs.CreateNamespace("mpas-system"),
 		envfuncs.CreateNamespace(namespace),
 		shared.StartGitServer(namespace),
