+++ internal/bootstrap/install_cert_manager.go
@@ -13,6 +13,7 @@ import (
 	"time"
 
 	"github.com/fluxcd/go-git-providers/gitprovider"
+	"github.com/open-component-model/mpas/internal/env"
 	"github.com/open-component-model/ocm/pkg/contexts/ocm"
 )
 
@@ -55,6 +56,7 @@ func newCertManagerInstall(name, version string, repository ocm.Repository, opts
 			version:       version,
 			repository:    repository,
 			dir:           opts.dir,
+			host:          env.DefaultCertManagerHost,
 		}),
 	}
